Refinements to uniformize the API with the plugis one.
