+++ errbot/__init__.py
@@ -13,9 +13,10 @@ from errbot.backends.base import Message, ONLINE, OFFLINE, AWAY, DND  # noqa
 from .utils import compat_str
 from .utils import PY2, PY3  # noqa gbin: this is now used by plugins
 from .botplugin import BotPlugin, SeparatorArgParser, ShlexArgParser  # noqa
+from .flow import Flow, BotFlow
 from .core_plugins.wsview import route, view  # noqa
 
-__all__ = ['BotPlugin', 'webhook', 'webroute', 'webview', 'botcmd', 're_botcmd', 'arg_botcmd']
+__all__ = ['BotPlugin', 'webhook', 'webroute', 'webview', 'botcmd', 're_botcmd', 'arg_botcmd', 'botflow', 'BotFlow', 'Flow']
 
 log = logging.getLogger(__name__)
 
@@ -402,6 +403,22 @@ def cmdfilter(*args, **kwargs):
     def decorate(func):
         if not hasattr(func, '_err_command_filter'):  # don't override generated functions
             func._err_command_filter = True
+        return func
+
+    if len(args):
+        return decorate(args[0], **kwargs)
+    return lambda func: decorate(func, **kwargs)
+
+
+def botflow(*args, **kwargs):
+    """
+    Decorator for flow of commands.
+
+    TODO(gbin): example / docs
+    """
+    def decorate(func):
+        if not hasattr(func, '_err_flow'):  # don't override generated functions
+            func._err_flow = True
         return func
 
     if len(args):
